refactor(grading): split composite.py into per-concern package (Stage 1 of 3)
- New composite/ package with per-concern modules: state_checks, transcript_rules, trace_checks, llm_judge, custom_checks
- composite/__init__.py re-exports 5 grade functions + build_judge_state_diff + StateChecksReadResult + _build_runner_check_transcript (canonical fixture usage) + transient _check_result_to_wire / _executor_error_to_wire (removed in Stage 2)
- Retarget test_grading_substrate_parity.py monkeypatches to sub-modules (LEGB: package re-export is a no-op for sub-module-internal callees; EVALUATED at state_checks, evaluate_trace_checks parametrize row at trace_checks)
- Retarget test_grading_ledger.py string-form module for the same LEGB reason (evaluate_trace_checks lives at composite.trace_checks post-split)
- test_grader_parity_reference.py:418 string monkeypatch preserved via package re-export
- Byte-identical (Stage 2 drops pb2; Stage 3 adds guardrail)

Refs #1351
